+++ src/std/alloc/src/lib.rs
@@ -1,4 +1,16 @@
-use std::{alloc::Layout, os::raw::c_int, ptr};
+use nstd_core::pointer::NSTDPointer;
+use std::{
+    alloc::Layout,
+    os::raw::{c_int, c_void},
+    ptr::addr_of_mut,
+};
+
+/// Represents a heap allocated object.
+#[repr(C)]
+pub struct NSTDHeap {
+    /// Raw pointer to heap allocated object.
+    pub ptr: NSTDPointer,
+}
 
 /// Allocates a new memory block.
 /// Parameters:
@@ -9,7 +21,7 @@ use std::{alloc::Layout, os::raw::c_int, ptr};
 pub unsafe extern "C" fn nstd_std_alloc_allocate(size: usize) -> *mut u8 {
     match Layout::array::<u8>(size) {
         Ok(layout) => std::alloc::alloc(layout),
-        _ => ptr::null_mut(),
+        _ => std::ptr::null_mut(),
     }
 }
 
@@ -22,7 +34,7 @@ pub unsafe extern "C" fn nstd_std_alloc_allocate(size: usize) -> *mut u8 {
 pub unsafe extern "C" fn nstd_std_alloc_allocate_zeroed(size: usize) -> *mut u8 {
     match Layout::array::<u8>(size) {
         Ok(layout) => std::alloc::alloc_zeroed(layout),
-        _ => ptr::null_mut(),
+        _ => std::ptr::null_mut(),
     }
 }
 
@@ -65,9 +77,37 @@ pub unsafe extern "C" fn nstd_std_alloc_deallocate(ptr: *mut *mut u8, size: usiz
     match Layout::array::<u8>(size) {
         Ok(layout) => {
             std::alloc::dealloc(*ptr, layout);
-            *ptr = ptr::null_mut();
+            *ptr = std::ptr::null_mut();
             0
         }
         _ => 1,
     }
+}
+
+/// Creates a new heap allocated object.
+/// Parameters:
+///     `const NSTDPointer *const ptr` - Pointer to an object to be copied to the heap.
+/// Returns: `NSTDHeap obj` - The new heap allocated object.
+#[cfg_attr(feature = "clib", no_mangle)]
+pub unsafe extern "C" fn nstd_std_alloc_heap_new(ptr: &NSTDPointer) -> NSTDHeap {
+    let ptr_slice = core::slice::from_raw_parts(ptr.ptr as *const u8, ptr.size);
+    let alloc = nstd_std_alloc_allocate(ptr.size);
+    if !alloc.is_null() {
+        let alloc_slice = core::slice::from_raw_parts_mut(alloc, ptr.size);
+        alloc_slice.copy_from_slice(ptr_slice);
+    }
+    NSTDHeap {
+        ptr: nstd_core::pointer::nstd_core_pointer_new(alloc as *mut c_void, ptr.size),
+    }
+}
+
+/// Frees a heap allocated object.
+/// Parameters:
+///     `NSTDHeap *const obj` - The heap allocated object.
+/// Returns: `int errc` - Nonzero on error.
+#[inline]
+#[cfg_attr(feature = "clib", no_mangle)]
+pub unsafe extern "C" fn nstd_std_alloc_heap_free(obj: &mut NSTDHeap) -> c_int {
+    let ptr = addr_of_mut!(obj.ptr.ptr) as *mut *mut u8;
+    nstd_std_alloc_deallocate(ptr, obj.ptr.size)
 }