+++ src/std/alloc/CHANGELOG.md
@@ -1,4 +1,6 @@
 # nstd_alloc
+## Latest
+- Added `NSTDHeap` type for single object heap allocation.
 ## 0.2.0
 - `clib` feature.
 ## 0.1.0